pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parallax-agents"
version = "0.1.0"
description = "Autonomous multi-agent UI workflow capture with Playwright"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{ name="Parallax" }]

dependencies = [
  "playwright>=1.45",
  "typer[all]>=0.9.0",
  "structlog>=24.1.0",
  "prometheus-client>=0.20.0",
  "pyyaml>=6.0.1",
  "httpx>=0.27.0",
  "pillow>=10.0.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["parallax*"]
exclude = ["tests*", "configs*"]

[project.optional-dependencies]
llm = [
  "openai>=1.40.0",
  "anthropic>=0.34.0",
]

[project.scripts]
parallax = "parallax.runner.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

